Create UserStorageReport.php
This file is run from a cron job to check if a user has gone over the storage size they selected when first subscribing for the storage service.

It gathers the status of each user into a report to email to an admin user, and provides the line numbers to comment out so that the report only contains errors.

Since this isn't intended to be an exact check that will prevent users from accessing any storage above the limit they've paid for, there is some freedom allowing users to access the service before seeing a warning about going over their limit. 

If I wanted to make this more exact to display a warning message, I could set a $margin_of_error variable at the discrepancy level the client is comfortable with, then associate this with a condition and a corresponding error and report text, if I wanted to display a different message warning users that they are about to go over their limit, and can upload data equivalent to the discrepancy.

Or I could run a storage check with every call to the storage service, to display a message about whether this transaction will put the user over their limit.

The cron job is in the comment at the end. The directory path to your PHP binary or the application server user may be different for you.